youtube/requirements.txt
requests
Flask
google-genai
python-dotenv
